requirements.txt
fastapi
uvicorn
langchain
chromadb
ollama
python-dotenv
pypdf
